requirements.txt
fastapi~=0.100.0
pymongo~=4.4.1
pytest~=7.4.0

SQLAlchemy~=2.0.19
pydantic~=2.0.3
python-dotenv~=1.0.0
alembic~=1.11.2
redis~=4.6.0
celery~=5.3.1
flower